Projeto de Sistemas 02_25/backend/backend/requirements.txt

        asgiref==3.6.0
        autopep8==1.5.7
        beautifulsoup4==4.11.1
        behave==1.2.6
        behave-django==1.4.0
        certifi==2021.10.8
        cffi==1.15.0
        charset-normalizer==2.0.7
        coreapi==2.3.3
        coreschema==0.0.4
        cryptography==35.0.0
        Deprecated==1.2.13
        dj-database-url==0.5.0
        Django==4.1.5
        django-cpf-cnpj==1.0.0
        django-easy-audit==1.3.3
        django-filter==21.1
        django-oauth-toolkit==1.5.0
        django-polymorphic==3.1.0
        djangorestframework==3.12.4
        drf-yasg==1.20.0
        gunicorn==20.1.0
        idna==3.3
        inflection==0.5.1
        itypes==1.2.0
        Jinja2==3.0.3
        jwcrypto==1.0
        Markdown==3.3.4
        MarkupSafe==2.0.1
        oauthlib==3.1.1
        packaging==21.2
        parse==1.19.0
        parse-type==0.5.2
        psycopg2==2.9.2
        pycodestyle==2.7.0
        pycparser==2.21
        pyparsing==2.4.7
        python-decouple==3.5
        pytz==2021.3
        requests==2.26.0
        rest-condition==1.0.3
        ruamel.yaml==0.17.17
        ruamel.yaml.clib==0.2.6
        six==1.16.0
        soupsieve==2.3.2.post1
        sqlparse==0.4.3
        toml==0.10.2
        Unipath==1.1
        uritemplate==4.1.1
        urllib3==1.26.7
        whitenoise==5.3.0
        wrapt==1.13.3
        # Imported from the grammar
        
    